proto-django-rest/django/requirements.txt
django~=2.2.7
djangorestframework~=3.13.1
django-cors-headers~=3.1.1

#drf-jwt~=1.13.3
drf-jwt~=1.19.2
python-dotenv~=0.10.3
requests~=2.28
urllib3~=1.26
cryptography~=3.4
#pyjwt~=1.7.1
PyJWT~=2.4.0
pyjwt-key-fetcher~=0.1.1 
py-eureka-client>=0.11.3
spring-config-client>=0.2 
